Move parameters in a single folder
Previously, parameters were put deep inside the code tree.
It's now set in a base config folder.

If it's relevant, the parameters.js file could be split in multiple ones.
Moreover, it can be a good idea to make it a node module in order to avoid
the `../../../../` effect in imports.